hacky accounting for difference between GMT and pacific time
