+++ scripts/nightly-update/package.json
@@ -11,5 +11,7 @@
   "devDependencies": {},
   "author": "John Gravois <jgravois@esri.com> (http://johngravois.com)",
   "license": "Apache-2.0",
-  "readmeFilename": "README.md"
+  "readmeFilename": "README.md",
+  "scripts": {
+  }
 }
